Refactoring: Put the pathfinding code on the server into separate files pathfinding.cpp/hpp to reduce the size of the Game class
