+++ src/FreeAge/server/pathfinding.hpp
@@ -1,0 +1,10 @@
+// Copyright 2020 The FreeAge authors
+// This file is part of FreeAge, licensed under the new BSD license.
+// See the COPYING file in the project root for the license text.
+
+#pragma once
+
+class ServerMap;
+class ServerUnit;
+
+void PlanUnitPath(ServerUnit* unit, ServerMap* map);
